Renamed module that LodeRunner class is under to loderunner, created a new bomberman module.
