+++ applications/harnesses/ch_DDP_loderunner/train_LodeRunner_ddp.py
@@ -7,7 +7,7 @@ import torch.nn as nn
 import torch.distributed as dist
 from torch.nn.parallel import DistributedDataParallel as DDP
 
-from yoke.models.vit.swin.bomberman import LodeRunner
+from yoke.models.vit.swin.loderunner import LodeRunner
 from yoke.datasets.lsc_dataset import LSC_rho2rho_temporal_DataSet
 from yoke.utils.training.epoch.loderunner import train_DDP_loderunner_epoch
 from yoke.utils.restart import continuation_setup
