+++ applications/harnesses/chicoma_lsc_loderunner-ch-subsampling/train_density_LodeRunner.py
@@ -19,7 +19,7 @@ import random
 import logging
 import numpy as np
 
-from yoke.models.vit.swin.bomberman import LodeRunner
+from yoke.models.vit.swin.loderunner import LodeRunner
 from yoke.datasets.lsc_dataset import LSC_rho2rho_temporal_DataSet
 from yoke.utils.training.epoch import train_LRsched_loderunner_epoch
 from yoke.utils.dataload import make_dataloader
